requirements.txt
django
django-ninja
django-ninja-jwt[crypto]
rav
